+++ README.md
@@ -1,4 +1,4 @@
-# iopy
+# iotables
 
 ## Description
 
@@ -14,7 +14,7 @@ Currently the following databases are supported:
 
 ## Installation
 ```bash
-pip install git+https://github.com/WWakker/iopy.git
+pip install iotables
 ```
 
 ## Structure of input-output data
@@ -176,16 +176,16 @@ Creating an instance of the OECD class loads the OECD data and gives access to c
 An instance can be create as follows, specifiying a year between 1995 and 2018.
 
 ```python
-import iopy 
-oecd = iopy.OECD(version='2021', year=2018, refresh=False)
+import iotables 
+oecd = iotables.OECD(version='2021', year=2018, refresh=False)
 ```
 
 Similary, an instance can be created for other data, for example Figaro and ExioBase as
 
 ```python
-import iopy 
-figaro = iopy.Figaro(version='2022', year=2020, kind='industry-by-industry')
-exio = iopy.ExioBase(version='3.81', year=2022, kind='industry-by-industry')
+import iotables 
+figaro = iotables.Figaro(version='2025', year=2020, kind='industry-by-industry')
+exio = iotables.ExioBase(version='3.81', year=2022, kind='industry-by-industry')
 ```
 
 Creating an instance of a database class downloads and loads the data into memory, creates standard input-output matrices, and gives access to the following attributes and methods:
@@ -223,9 +223,9 @@ All matrices are extended `numpy.ndarray`'s with attributes `info`, `rows` and `
 
 When running a Leontief or Ghosh shock, the percentage shock to final demand/primary inputs in countries and sectors can be specified as
 ```python
-import iopy
+import iotables
 
-oecd = iopy.OECD(version='2021', year=2018)
+oecd = iotables.OECD(version='2021', year=2018)
 
 df_l = oecd.leontief_demand_shock(shock=-10, regions=['FR', 'DE], sectors=['01T02', '35'])
 df_g = oecd.ghosh_supply_shock(shock=-10, regions=['FR', 'DE'], sectors=['01T02', '35'])
