--- iopy/__about__.py
@@ -1,6 +1,0 @@
-__title__ = "iopy"
-__about__ = "A Python package to easily load and work with input-output data"
-__version__ = '0.0.1'
-__authors__ = 'Wouter Wakker'
-__email__ = "wouter.wakker@outlook.com"
-__url__ = ""
