--- iopy/core/__init__.py
@@ -1,7 +1,0 @@
-"""  Created on 22/11/2022::
-------------- __init__.py -------------
-**Authors**: W. Wakker
-
-"""
-
-from iopy.core.globals import *
