--- iopy/dev-requirements.txt
@@ -1,8 +1,0 @@
-requests
-curl_cffi
-numpy
-pandas>=1.1.5
-matplotlib
-tqdm
-pytest
-pytest-cov
