--- iopy/requirements.txt
@@ -1,6 +1,0 @@
-requests
-curl_cffi
-numpy
-pandas>=1.1.5
-matplotlib
-tqdm
