--- iopy/tests/.coveragerc
@@ -1,5 +1,0 @@
-[run]
-omit = res/*
-       **/__about__.py
-       **/adjust_source_in_covxml.py
-       **/update_badge.py
