+++ test.py
@@ -1,3 +1,3 @@
-from iopy import OECD
+from iotables import OECD
 
 io = OECD(version='2025-extended', year=2022, refresh=True)
